+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Extracted typed response mappings to telegram-bot-types gem.
   It now provides definitions for all API v3.5 methods.
+- Fix #chat for non-message updates with TypedUpdates enabled
 
 # 0.13.0